src/__main__.py
import asyncio

from discord import Intents, Message
from discord.ext.commands import Bot, Cog
from loguru import logger
from src.settings import settings

intents = Intents.default()
intents.message_content = True

bot = Bot(command_prefix="ai!", intents=intents)


class BotCog(Cog):
    def __init__(self, client: Bot):
        self.client = client

    @Cog.listener()
    async def on_ready(self):
        logger.info("Hello, Discord!")

    @Cog.listener()
    async def on_message(self, message: Message):
        if message.author != self.client.user:
            logger.info(f"{message.author.name}: {message.content}")


async def main():
    await bot.add_cog(BotCog(bot))
    await bot.start(settings.DISCORD_TOKEN)


if __name__ == "__main__":
    asyncio.run(main())
